Fix video with audio when using FFmpeg > 6.1
Since FFmpeg 6.1 the video generation for videos with audio is not
working as expected.
When generating the video there is an error: Queue input is backward in
time
Apparently the combination of atrim and adelay is not working very well
when a video with sound is included in the slideshow.

Instead of adelay silence is added to offset the audio of videos.

In the longterm it would be better to handle the audio track of videos
like the background audio so that the audio track is concated to the
background music sections.