Fix background music while videos with sound are playing
The background music was silently playing while a video with sound was
played. This leads to missing parts of the background music.
The music didn't resume where it was stopped when the video started.
Additionally the total duration of the background music was wrong
calculated.

This is fixed now by applying silence like for the videos in commit
https://github.com/Trekky12/kburns-slideshow/commit/c0ea6781746a989035063e2fe9480016eb671e11